design

+++ app/Http/Controllers/Product_commentsController.php
@@ -11,36 +11,32 @@ class Product_commentsController extends Controller
 {
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'content' => 'required',
-        ]);
-        
-        Product_comment::create([
-            'content' => $request->content,
-            'product_id' => $request-> product_id,
-            'user_id' => $request->user()->id,
-        ]);
-        
-        return back();
+        if(\Auth::check()) {
+            $this->validate($request, [
+                'content' => 'required',
+            ]);
+            
+            Product_comment::create([
+                'content' => $request->content,
+                'product_id' => $request-> product_id,
+                'user_id' => $request->user()->id,
+            ]);
+            
+            return back();
+        }
     }
     
     public function destroy($id)
     {
-        $delete_comment = Product_comment::find($id);
-        $product = Product::find($delete_comment->product_id);
-        $leader = User::find($product->leader_name);
-        $delete_comment->delete();
-        
-        $product_comments = $product->product_comments()->orderBy('created_at','desc')->paginate(10);
-        
-        
-        
-        $data = [
-            'product'=>$product,
-            'leader'=>$leader,
-            'product_comments'=>$product_comments,
-        ];
-        
-        return back();
+        if(\Auth::check()) {
+            $delete_comment = Product_comment::find($id);
+            $product = Product::find($delete_comment->product_id);
+            
+            if(\Auth::id() == $delete_comment->user_id) {
+                $delete_comment->delete();
+            }
+            
+            return back();
+        }
     }
 }
